backend/requirements.txt
axios
pydantic==2.11.1
fastapi
uvicorn
python-multipart
google-auth-oauthlib
google-auth-httplib2
google-api-python-client
websockets
python-dotenv
supabase
python-jose
